engdocs2llm/requirements.txt
flask>=3.0
# flask-cors>=4.0  # optional, only needed if running UI on different port
ezdxf>=1.3
pdfplumber>=0.11
pdf2image>=1.17
Pillow>=10.0
openpyxl>=3.1
python-docx>=1.1
anthropic>=0.40
networkx>=3.2
python-dotenv>=1.0
werkzeug>=3.0
